+++ packages/jest-each/src/__tests__/array.test.ts
@@ -291,6 +291,83 @@ describe('jest-each', () => {
           10000,
         );
       });
+
+      test('calls global with title containing object property when using $variable', () => {
+        const globalTestMocks = getGlobalTestMocks();
+        const eachObject = each.withGlobal(globalTestMocks)([
+          {
+            a: 'hello',
+            b: 1,
+            c: null,
+            d: undefined,
+            e: 1.2,
+            f: {key: 'foo'},
+            g: () => {},
+            h: [],
+            i: Infinity,
+            j: NaN,
+          },
+          {
+            a: 'world',
+            b: 1,
+            c: null,
+            d: undefined,
+            e: 1.2,
+            f: {key: 'bar'},
+            g: () => {},
+            h: [],
+            i: Infinity,
+            j: NaN,
+          },
+        ]);
+        const testFunction = get(eachObject, keyPath);
+        testFunction(
+          'expected string: %% %%s $a $b $c $d $e $f $f.key $g $h $i $j $#',
+          noop,
+        );
+
+        const globalMock = get(globalTestMocks, keyPath);
+        expect(globalMock).toHaveBeenCalledTimes(2);
+        expect(globalMock).toHaveBeenCalledWith(
+          'expected string: % %s hello 1 null undefined 1.2 {"key": "foo"} foo [Function g] [] Infinity NaN 0',
+          expectFunction,
+          undefined,
+        );
+        expect(globalMock).toHaveBeenCalledWith(
+          'expected string: % %s world 1 null undefined 1.2 {"key": "bar"} bar [Function g] [] Infinity NaN 1',
+          expectFunction,
+          undefined,
+        );
+      });
+
+      test('calls global with title containing param values when using both % placeholder and $variable', () => {
+        const globalTestMocks = getGlobalTestMocks();
+        const eachObject = each.withGlobal(globalTestMocks)([
+          {
+            a: 'hello',
+            b: 1,
+          },
+          {
+            a: 'world',
+            b: 1,
+          },
+        ]);
+        const testFunction = get(eachObject, keyPath);
+        testFunction('expected string: %p %# $a $b $#', noop);
+
+        const globalMock = get(globalTestMocks, keyPath);
+        expect(globalMock).toHaveBeenCalledTimes(2);
+        expect(globalMock).toHaveBeenCalledWith(
+          'expected string: {"a": "hello", "b": 1} 0 $a $b $#',
+          expectFunction,
+          undefined,
+        );
+        expect(globalMock).toHaveBeenCalledWith(
+          'expected string: {"a": "world", "b": 1} 1 $a $b $#',
+          expectFunction,
+          undefined,
+        );
+      });
     });
   });